Make configurable the behavior of jump-to-test/impl
This is achieved by making the functions clojure-test-for and
clojure-test-implementation-for return the full path of the test or
src file. Next clojure-jump-to-test and
clojure-test-jump-to-implementation simply calls these functions
through a defvar to get the path. That var can be changed by users who
want a different implementation.

The existing behavior of where test files are create was left
unchanged.