Replace car and cdr with CL equivalents
`car` and `cdr` are both ancient and cryptic.  Since we're using cl
anyways it's a much better idea to use `first` and `rest` (not to
mention they'll be more familiar to Clojure hackers). `first` and `rest`
were already being used here and there, I've just wrapped the transition
to them.