Makes clojure-test-maybe-enable more discriminating.
It will only enable clojure-test-mode if it finds clojure.test somewhere in the
file.  This will prevent clojure-test-mode getting enabled for namespaces with
'test.' in their name that are for other testing frameworks.

It will also allow clojure-test-mode to co-exist with emacs modes for those
other frameworks.  (I thinking specifically of conjecture-mode.)