Do not add space between reader-literal tag and delimiter
This a stylistic choice.

Since the introduction of tagged reader literals, some people have
started writing code like this:

    #db/id[:db.part/user]

with no space between the tag and the opening delimiter.

This is particularly common with [Datomic], where the tag serves as a
pseudo-function-call in [EDN] files. For example, see
[seattle-schema.edn].

The only place I have never seen this is with string delimiters, for
example the built-in `#inst` tag, where I always use a space.

    #inst "2013-06-21T20:09:41.718-00:00"

Right now, clojure-mode will automatically insert a space between a
tag and a delimiter. I would like it to not insert that space, unless
the delimiter is `"`.