Separate symbols for namespace and vars
By changing the syntax class of / to punctuation, Emacs considers the
namespace part and var part of `namespace/var` as two separate symbols.

One improvement is that you can now hippie-expand namespaces and vars
separately. Another is that you can use symbol boundaries to match
namespaces and vars in regexen, with \\_< and \\_>