[Fix #343] Don't font-lock the `:` in keywords
The `:` in keywords were font-locked with the keyword face because we
inherited the rules for font-locking keywords from Elisp.

In Elisp a keyword is a symbol which starts with the `:` character.
Furthermore, the predicates symbolp and keywordp return T for keywords
in Elisp.  This is different from how a keyword is defined, and how it
behaves, in Common-lisp and in Clojure.

In common-lisp the `:` is a hint to the reader that the symbol should be
interned in the keyword package.

Clojure doesn't have a keyword package, but when the reader consumes a
`:` it means that the following symbol is to be stored in the keyword
table, residing in `Keyword.java`.  Unlike in Elisp, the `:` is never
stored anywhere. It's just the way keywords print and syntactic sugar
for keyword creation.

Thus, unlike in elisp, the `:` is a token meant for the reader, and not
part of the keyword itself, and should not be font-locked with the
keyword face.