[Fix 489] Inserting parens before comment form doesn't move point
In a form like

|

(comment
   (stuff))

Entering parens with paredit would put the parens right before the comment
block. Paredit determines if it is in a comment to insert parens so it doesn't
automatically enter a closing when in a comment or a string. Part of this called
beginning-of-defun which we have modified.

The error here was that rater than just going to the beginning of the form, we
went to the end and then back one logical form to be at the beginning. This is
identical behavior _unless_ you are between two forms. Going straight to the
beginning put you in the first form, going to the end and then the beginning
puts you in the second form. I.e.,

(formA)
|
(formB)

Our beginning of form went to formB but it should go to formA.