Handle unbalanced parens in custom beginning of defun fn
Paredit has some strange logic. It inserts its opening "(" and then tasks itself
with discovering whether it is in a comment or string before inserting its
closing ")". But this means that paredit _by design_ calls functions on
unbalanced parens states and I don't know why they do this.

If we find a scan-error, this bails and lets the default beginning of defun
function shoulder that logic.