[#422] Do not check for presence of `paredit-open-curly` and `paredit-close-curly` functions.
Passing non-nil fileonly parameter to check-declare to skip checking that the
functions are actually defined, because they are defined via the
`define-paredit-pair` macro.

See
https://www.gnu.org/software/emacs/manual/html_node/elisp/Declaring-Functions.html

> There are some function definitions that ‘check-declare’ does not
understand (e.g., defstruct and some other macros). In such cases, you can pass
a non-nil fileonly argument to declare-function, meaning to only check that the
file exists, not that it actually defines the function.