Support multiple consecutive reader comments (#_#_a b)
Modified the default regexps and the heuristic to find the end of the region to
comment out.

Previously Emacs would treat the second `#_` as the commented form, and then
highlight the following two forms as usual. Now it (mostly) matches what Clojure
actually evaluates. Things get weird when you start mixing `#_` and forms, but
this fixes the most common use cases, like key-value-pairs.