[Fix #551] Indent `clojure-align` region before aligning.
This prevents issues when aligning improperly indented code. Before, the region
would be aligned while retaining the improper indentation. The follow-up
indentation step would then break the alignment that was just performed.

With this commit, we switch the order by indenting first then aligning. This
retains compatibility with the test cases in #360.