Don't use custom logic to find source directory of clojure-mode.el
* Or this will prevent buttercup tests from working with installed clojure-mode
under ELPA directories.

* A bit more background: in Debian there is autopkgtest which tests the
installed package.  dh-elpa enables a mode that runs the same ERT or Buttercup
tests against the installed package instead of the source tree.  In this case,
it will let the tests pass during build phase, but autopkgtest will fail as this
custom code will still try to locate the source code which will no longer be
available.

* I have tested under Debian sbuild that removing the code will let
the buttercup tests pass under autopkgtest, but not sure whether this is still
required for Eldev to work.